RPC: Refactor linux RPC into platform.
Move common linux Rpc.* files into common platform to match refactoring
done for other platforms.
