+++ examples/platform/linux/Rpc.h
@@ -18,12 +18,10 @@
 
 #pragma once
 
-#include <CommonRpc.h>
-
 namespace chip {
 namespace rpc {
 
-class LightingService;
+int Init();
 
 } // namespace rpc
 } // namespace chip
